Revert "calculate a real range for the diagnostics"
This reverts commit 14fcf016b86ad67bad11be8b2ec2c6d127cf0b47.

After thinking about this more, we actually need to omit the range because that
will eventually mark a file-level diagnostic and suppress the snippet. The only
reason we're able to render the diagnostics shown in this commit is because
we're both setting a range, which is used in the header, *and* setting
`Annotation::is_file_level`, which suppresses the snippet rendering.

For example, setting `is_file_level` to `false` yields a diagnostic like this,
which I think we definitely don't want:

```
unformatted: File would be reformatted
 --> input.py:3:16
  |
1 |   from test import say_hy
2 |
3 |   if __name__ == '__main__':
  |  ________________^
4 | |     say_hy('dear Ruff contributor')
  | |_________________________________^
  |
1 | from test import say_hy
2 |
  - if __name__ == '__main__':
  -     say_hy('dear Ruff contributor')
3 + if __name__ == "__main__":
4 +     say_hy("dear Ruff contributor")

1 file would be reformatted
```