emit a warning for non-preview --output-format usage
as a global option, preview comes from the resolved settings, so I just moved
the pyproject.toml resolution outside of `format` and `format_stdin` since it
was called immediately in both cases anyway.

note that this won't warn for usage of the `output-format` setting within a
pyproject.toml because that option will already have been `unwrap_or_default`ed
by the time we could check it here.